# contracts/taskflow/openapi.yaml
openapi: 3.1.0
info:
  title: TaskFlow Learning Contract
  version: 1.0.0
servers:
  - url: http://127.0.0.1:3000
paths:
  /health:
    get:
      operationId: getHealth
      responses:
        "200":
          description: Service is running
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status: {type: string, const: ok}
  /tasks:
    get:
      operationId: listTasks
      responses:
        "200":
          description: Task collection
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: {$ref: "#/components/schemas/Task"}
    post:
      operationId: createTask
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: {type: string, minLength: 1, maxLength: 128}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CreateTask"}
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Task"}
        "200":
          description: Existing result for repeated idempotency key
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Task"}
        "400": {$ref: "#/components/responses/Problem"}
        "415": {$ref: "#/components/responses/Problem"}
        "422": {$ref: "#/components/responses/Problem"}
  /tasks/{taskId}:
    get:
      operationId: getTask
      parameters:
        - name: taskId
          in: path
          required: true
          schema: {type: string}
      responses:
        "200":
          description: Task found
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Task"}
        "404":
          description: TASK_NOT_FOUND
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Problem"}
              example:
                error: {code: TASK_NOT_FOUND, message: Task was not found}
components:
  schemas:
    CreateTask:
      type: object
      additionalProperties: false
      required: [title]
      properties:
        title: {type: string, minLength: 1, maxLength: 120}
    Task:
      type: object
      required: [id, title, completed, createdAt]
      properties:
        id: {type: string}
        title: {type: string}
        completed: {type: boolean}
        createdAt: {type: string, format: date-time}
    Problem:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: {type: string}
            message: {type: string}
  responses:
    Problem:
      description: Stable error envelope
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Problem"}